Move EntryPoint and GameRoot to Controller folder
